changing services name and moving getData to services
